+++ update/name.go
@@ -1,12 +1,5 @@
 package update
 
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-)
-
 var nameCache = make(map[string]string)
 
 func GetName(manga string) (string, error) {
@@ -20,25 +13,7 @@ func GetName(manga string) (string, error) {
 		return seriesJSON.Metadata.Name, nil
 	}
 
-	// recursively search for .cbz files
-	// find the first one and get the name from it
-	var cbzFiles []string
-	err = filepath.Walk(manga, func(path string, info os.FileInfo, err error) error {
-		if strings.HasSuffix(path, ".cbz") {
-			cbzFiles = append(cbzFiles, path)
-		}
-		return nil
-	})
-
-	if err != nil {
-		return "", err
-	}
-
-	if len(cbzFiles) == 0 {
-		return "", fmt.Errorf("no .cbz files found")
-	}
-
-	comicInfo, err := getComicInfoXML(cbzFiles[0])
+	comicInfo, err := getAnyChapterComicInfo(manga)
 	if err != nil {
 		return "", err
 	}