+++ CHANGELOG.md
@@ -47,6 +47,8 @@ See [Inline mode wiki](https://github.com/metafates/mangal/wiki/Inline-mode) for
 - Improved TUI experience
 - Search completions in TUI. `mangal config info -k search.show_query_suggestions`
 - Anilist caching significantly improved. Now, it will cache all search results (for 2 days)
+- Update metadata of already downloaded manga (ComicInfo.xml, series.json, cover image) after changing Anilist bind. #124
+  See `mangal inline anilist update`
 - Add caching for custom (lua) sources
 - Include different cover sizes and color for json output #116
 - Add option to omit dates for ComicInfo.xml #117